feat: updating manager page
